Added Menu to the console

+++ navigation/graph.h
@@ -5,12 +5,12 @@ class graph
 {
 public:
 	int V, E;
-	std::vector<std::pair<int, ipair>> edges;
+	std::vector < std::pair < double , ipair >> edges;
 	std::vector<int> dfs[11];
 	bool visited[11];
 	graph(int ver=0 , int edg=0);
-	void addedge(int u, int v, int w);
-	int kruskalMST();
+	void addedge(int u, int v, double w);
+	double kruskalMST();
 	void caller(int s, int d);
 	int dfsf(int st, int dest);
 	~graph();
